feat: searching user in empPage

+++ src/main/java/edu/example/bts/service/JwtService.java
@@ -17,7 +17,7 @@ public class JwtService {
 	private static final SecretKey SECRET_KEY = Keys.secretKeyFor(SignatureAlgorithm.HS256);
 	private static final long EXPIRATION_TIME = 1000L * 60 * 30; // 30분
 
-	public String createToken(String email, int empno, String role) {
+	public String createToken(String email, Long empno, String role) {
 		return Jwts.builder().setSubject(email).claim("empno", empno).claim("role", role).setIssuedAt(new Date())
 				.setExpiration(new Date(System.currentTimeMillis() + EXPIRATION_TIME))
 				.signWith(SECRET_KEY, SignatureAlgorithm.HS256)
